[m3admin] make logger + http client configurable (#20)
While working on cluster resizing I needed to dump the raw response from
the coordinator in order to view the underlying error. I thought it
might be easier to add that by default when the logger is set to debug,
and thought it would be easier to make the m3admin http client an
interface paramaterized by its logger and http client. The configurable
http client is for the purpose of tests where we don't want to wait the
default retry times for unit tests to complete.

The change to expose the m3cluster `Placement` is motivated by another
change as part of cluster resizing that was easier to just pull in now.