[tests] add e2e integration framework + tests
This adds a basic integration test for ensuring operator functionality
against a real Kubernetes cluster.

So far only basic functionality is tested: installing the operator and
creating a 3 node M3DB cluster with no persistent storage. The tests
ensure the cluster is created within a certain amount of time and that
the placement is valid.

Note that these tests currently do not run on CI, but such functionality
is planned. Eventually we'll have creds and infrastructure to create a
GKE cluster from CI on certain commits. Right now the tests have to be
run locally against a working Kubernetes cluster.