[api] make pod identity configurable
We need a manner of determining pod uniqueness, as pod names remain the
same across restarts, rescheduling of storage, etc. This PR introduces
the concept of "pod identity", which is created from a configurable list
of sources to determine the uniqueness of a pod.

This adds a new controller that watches pods, and if they're part of an
M3DB cluster annotates them with their identity. This annotation is
passed to the pod through a downward API volume and picked up by dbnode
instances when they start up and look for themselves in the placement.
All placement operations are based off of this identity.