[api] represent durations as strings (#94)
We'd previously been using `time.Duration` as the type of various
durations in the namespace options. I swear this was working a while ago
so I'm not sure if something changes with `client-go` or if I totally
missed and issue. This changes the CRD to use strings and parse the
durations separately, which should fix the issue of the string
durations failing to parse.

Also added some context to the namespace create error and include it in
an event.