[api] support per-isogroup storageclasses (#98)
Users running Kubernetes < 1.12, in which topology-aware volume
scheduling was introduced, cannot use a single storageclass for all
their statefulsets as there is no guarantee that the PV for a pod will
be created in the zone the pod is pinned to. This can cause pods to be
stuck in an unschedulable state.

This allows users to specify a StorageClass per-statefulset (zone). This
solves the above problem as a single StorageClass can be created with
attributes to specify that PVs created with it will be in a given zone
(see
https://v1-11.docs.kubernetes.io/docs/concepts/storage/storage-classes/).