[api] cluster spec etcd endpts; per-cluster envs (#99)
This PR has a few changes to improve the interaction between
operator-operated M3DB clusters and etcd:

1. Users can now specify etcd endpoints on the m3dbcluster spec,
removing one of the most common changes users would have to make in
their config files.

2. The operator will include a per-cluster environment header in all
requests to the coordinator, meaning that namespaces and placements will
be stored under separate keys per-cluster in etcd and multiple
m3dbclusters can share the same etcd cluster.

**NOTE**: This PR introduces a breaking change that will require users
to 1) rename 2 keys in their etcd clusters and 2) change a value in
their configmap if they would like the operator to continue being able
to manage clusters from before this commit. The changelog for this
release will include instructions for this.