[api] support custom node affinity per-isogroup (#106)
Previously we had been using the name of an IsolationGroup as its
the value selector in a StatefulSet's NodeAffinity, with the key
hardcoded to `failure-domain.beta.kubernetes.io/zone`. This meant that
users had to have clusters across 3 unique zones.

This relaxes those constraints and allows users to specify the
NodeAffinity selectors they wish to use, meaning they can have clusters
within a single zone or other failure domain.