[scripts] add script for migrating etcd data
In the next release of the operator we'll be changing the default m3db
configmap, and thus how data is stored in etcd. This provides a script
that will let users migrate their data from one etcd keyspace to another
based on their cluster name.

Instructions for using the script to be included in release notes.