[m3admin] don't set per-environment header (#110)
We added per-environment headers to support the case where a user was
changing the environment and we didn't want a race based on which nodes
got the new config first. However, this ends up adding more
complications. Let's assume that environments will always stay the same
(if using the default configmap we never change it) and make this logic
easier.