[m3admin] restore per-environment header
This reverts commit 45e6abb4acc9edb6a18d04ffe449956c2d2b8b19.

We made that change under the assumption that m3coordinator would use
the environment in its config file as the default if none was set, but
it turns out it always defaults to `default_env`. Instead we'll add
warning to the documentation that users must always set their env to
`$NAMESPACE/$CLUSTER` if using a custom config.