[controller] continuously reconcile configmaps (#118)
Now that we have a default template for configmaps that we construct for
users, we should continuously reconcile it. This will allow us to roll
out new config for M3DB over time and know that users' clusters will
pick up our changes.

This logic doesn't trigger if the user has a custom configmap.