[api] expand node affinity; support multiple reqs
Previously we supported node affinity by either 1) using the default
heuristic of the isolation group name being the zone, or 2) by
specifying a single key and a set of values to match an m3db pod to a
node.

This limits users from specifying more complex terms such as "nodes in
zone A with instance type 'large'".

This change allows users to specify multiple node affinity terms, as
well as specifying none at all if they have no need for node affinity.
This is a breaking change as users must specify the zone key in their
cluster configs to get the same zone affinity as before, and this will
be called out in the release notes.

Eventually we'll add the ability to specify entire affinity terms per
isolation-group, but those can be extremely verbose and this still
provides a nice shortcut for most use cases.