Validate placement before reconciling namespaces
Marking namespaces as ready requires the coordinator to check
each dbnode to see if the namespace is available (via a read).
This cannot proceed successfully without a placement as the
dbnode blocks until it has one. Therefore, ensure placement
gets created before namespaces so dbnode can respond to
the coordinator.