[controller] Ensure k8s statefulset statuses are fresh (#271)
In debugging #268, I discovered that in all cases where we updated more
than 1 statefulset, it was because the observed generation on the set
was 1 less than the set's actual generation. This means that the
Kubernetes statefulset controller had not yet processed the update, and
that the information on `Status` that we check (ready replicas, current
replicas, etc.) was not up-to-date.

Also out of paranoia, I changed the update behavior to ensure that we
keep the old object meta fields around when we send the `Update` to
Kubernetes. I think if this isn't the case, we potentially overwrite a
set even if it had changed since we last process it.

This PR ensures that we stop processing the cluster if any of the set
statuses are not fresh, and that we use familiar conflict methods on set
updated.

Closes #268.